Enemy Fix, Knockback (E and P), State Packets, Enemy Animations
States send

+++ UTR - Server/Scripts/Items/Tool.cs
@@ -11,10 +11,10 @@ namespace items
 		[Export]
 		public int baseDmg;
 		[Export]
-		public int lvlScale;
+		public float lvlScale;
 		[Export]
 		public int range;
 		[Export]
-		public int cooldown;
+		public float cooldown;
 	}
 }
